feat: autorefresh the empty contact list
